task five: 100-elf_header.c - status: edit - desc: fix bugs.

+++ 0x15-file_io/main.h
@@ -34,7 +34,7 @@ typedef struct ElfABI
 
 typedef struct ElfType
 {
-	unsigned long int typeByte;
+	unsigned int typeByte;
 	char typeString[100];
 } elfType;
